# avoid repeated large allocations for large parts